chore: Rename server module to ports
