+++ internal/ports/errors.go
@@ -1,4 +1,4 @@
-package server
+package ports
 
 import (
 	"context"
