+++ internal/ports/errors_test.go
@@ -1,4 +1,4 @@
-package server
+package ports
 
 import (
 	"context"
