+++ internal/ports/middleware_test.go
@@ -1,4 +1,4 @@
-package server
+package ports
 
 import (
 	"net/http"
