+++ internal/ports/player_data.go
@@ -1,4 +1,4 @@
-package server
+package ports
 
 import (
 	"context"
