+++ internal/ports/player_data_test.go
@@ -1,4 +1,4 @@
-package server
+package ports
 
 import (
 	"context"
